Support latest version of pack
Additionally, added a check that warns if the installed
version of pack is different than what's expected (typically
latest).

We were seeing failures in CI that we weren't seeing locally.
It came down to CI running a newer (latest) version of pack
in which the platform API was changed so therefore the output
changed slightly.

Signed-off-by: Javier Romero <rjavier@vmware.com>